feat: links in home page and rover pages now work
All except the info link.
One the home page when a rover link is clicked then page is re-render
to the rover page, when this occurs the nav elements are re-drawn
this means that the event listeners associated with it on the home
page are lost and they have to be re-established. To achieve this,
when the application store is updated the updateStore function
checks to see if it was the page has changed, if it has it calls
a function to create the event listeners for the new page.
Note this is the same for all the rover links on the home page as well.