feat: front end shows latest images
Set up a new api request from the front to the backend that
requests the latest images from the nasa rover api.
'latest' is fetched from the front end, the backend has a new
endpoint that deals with this and new functions to fetch and
request data from the nase api. A list of images and rover info,
is returned to the front end and display on the rover page